moved auth context > frontend, add firebase env
