+++ backend/firebase/firebase.tsx
@@ -6,23 +6,18 @@ import { addDoc, collection, getDocs, getFirestore, query, where } from "firebas
 // TODO: Add SDKs for Firebase products that you want to use
 // https://firebase.google.com/docs/web/setup#available-libraries
 
-// Your web app's Firebase configuration
-// For Firebase JS SDK v7.20.0 and later, measurementId is optional
 const firebaseConfig = {
-  apiKey: "AIzaSyAWCKEkLBpYgN_dIJl6ncEHPngL21j3kz4",
-  authDomain: "okb-hope.firebaseapp.com",
-  projectId: "okb-hope",
-  storageBucket: "okb-hope.appspot.com",
-  messagingSenderId: "321787147976",
-  appId: "1:321787147976:web:faae136e03900bd47b97fd",
-  measurementId: "G-V7SV4XK6GC"
+  apiKey: process.env.NEXT_PUBLIC_FIREBASE_API_KEY,
+  authDomain: process.env.NEXT_PUBLIC_FIREBASE_AUTH_DOMAIN,
+  projectId: process.env.NEXT_PUBLIC_FIREBASE_PROJECT_ID,
+  storageBucket: process.env.NEXT_PUBLIC_FIREBASE_STORAGE_BUCKET,
+  messagingSenderId: process.env.NEXT_PUBLIC_FIREBASE_MESSAGING_SENDER_ID,
+  appId: process.env.NEXT_PUBLIC_FIREBASE_APP_ID,
+  measurementId: process.env.NEXT_PUBLIC_FIREBASE_MEASUREMENT_ID
 };
 
 // Initialize Firebase
 const app = initializeApp(firebaseConfig);
-// if (typeof window !== "undefined") {
-//   const analytics = getAnalytics(app);
-// }
 
 const analytics = app.name && typeof window !== 'undefined' ? getAnalytics(app) : null;
 
@@ -57,7 +52,7 @@ const signInWithGoogle = async () => {
   catch (err) {
     console.error(err);
     if (err instanceof Error) {
-      alert(err.message);
+      alert("An error occurred while signing in with Google: " + err.message);
     }
   }
 };
