--- backend/temp.tsx
@@ -1,1 +1,0 @@
-// hello
