+++ frontend/src/pages/_app.tsx
@@ -1,7 +1,6 @@
 import '@/styles/globals.css'
 import type { AppProps } from 'next/app'
-import { AuthProvider } from '../../../backend/firebase/AuthContext'
-
+import { AuthProvider } from '../../contexts/AuthContext'
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
